# api_specification/openapi.yml
openapi: '3.0.2'
info:
  title: 'API Title'
  version: '1.0'
servers:
  - url: https://api.server.test/v1
paths:
...
  /candidate:
    post:
      summary: Create a candidate
      description: Creates a candidate
      operationId: createCandidate
      requestBody:
        description: Candidate
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /candidate/{id}:
    get:
      summary: Get a candidate
      description: Returns a candidate
      operationId: getCandidate
      parameters:
        - name: id
          in: path
          description: Candidate ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
    put:
      summary: Update a candidate
      description: Updates a candidate
      operationId: updateCandidate
      parameters:
        - name: id
          in: path
          description: Candidate ID
          required: true
          schema:
            type: string
      requestBody:
        description: Candidate
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a candidate
      description: Deletes a candidate
      operationId: deleteCandidate
      parameters:
        - name: id
          in: path
          description: Candidate ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...
  schemas:
    Candidate:
      type: object
      required:
        - id
        - name
        - description
        - createdAt
        - updatedAt
        - deletedAt
        - roles
        - qualifications
      properties:
        id:
          type: string
        name:
          type: string
          description:
            The name of the candidate
        description:
          type: string
          description:
            The description of the candidate
        createdAt:
          type: string
          format: date-time
          description:
            The date and time when the candidate was created
        updatedAt:
          type: string
          format: date-time
          description:
            The date and time when the candidate was updated
        deletedAt:
          type: string
          format: date-time
          description:
            The date and time when the candidate was deleted
        roles:
          description:
            The roles of the candidate
          type: array
          items:
            $ref: '#/components/schemas/Role'
        qualifications:
          description:
            The qualifications of the candidate
          type: array
          items:
            $ref: '#/components/schemas/Qualification'
    Role:
      type: object
      required:
        - id
        - name
        - description
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
        name:
          type: string
          description:
            The name of the role
        description:
          type: string
          description:
            The description of the role
        createdAt:
          type: string
          format: date-time
          description:
            The date and time when the role was created
        updatedAt:
          type: string
          format: date-time
          description:
            The date and time when the role was updated
        deletedAt:
          type: string
          format: date-time
          description:
            The date and time when the role was deleted
        elected:
          type: boolean
          description:
            Whether a candidate has been elected to the role
        desired_qualifications:
          description:
            The desired qualifications of the role
          type: array
          items:
            $ref: '#/components/schemas/Qualification'
    Qualification:
      type: object
      required:
        - id
        - name
        - description
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
        name:
          type: string
          description:
            The name of the qualification
        description:
          type: string
          description:
            The description of the qualification
        createdAt:
          type: string
          format: date-time
          description:
            The date and time when the qualification was created
        updatedAt:
          type: string
          format: date-time
          description:
            The date and time when the qualification was updated
        deletedAt:
          type: string
          format: date-time
          description:
            The date and time when the qualification was deleted
security:
  - bearerAuth: []         # use the same name as above